tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/context.unrecognised.badsuffix.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: context.unrecognised.badsuffix</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>context.unrecognised.badsuffix</h1>
<p class="desc">Context name "2d" plus a suffix is unrecognised</p>


<script>
var t = async_test("Context name \"2d\" plus a suffix is unrecognised");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

var offscreenCanvas2 = new OffscreenCanvas(100, 50);
assert_throws(new TypeError(), function() { offscreenCanvas2.getContext("2d#"); });

t.done();

});
</script>
